// src/types/digiflazz.ts
export interface DigiflazzCheckSaldoResponse {
  data: {
    deposit: number
  }
}

export interface DigiflazzPriceListResponse {
  data: {
    product_name: string
    category: string
    brand: string
    type: string
    seller_name: string
    price: number
    buyer_sku_code: string
    buyer_product_status: boolean
    seller_product_status: boolean
    unlimited_stock: boolean
    stock: number
    multi: boolean
    start_cut_off: string
    end_cut_off: string
    desc: string
  }[]
}